demo/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-debug-toolbar==5.0.1
django-dump-die==0.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
setuptools==69.5.1
sqlparse==0.5.3
wheel==0.43.0
